Deduplicate Pumpkin/Melon stem code
